Use Decimal for Oracle float datatype

+++ QueryDB.Core.Tests/Entities.cs
@@ -147,7 +147,7 @@ namespace QueryDB.Core.Tests
                 public string Char_Column { get; set; }
                 public string Clob_Column { get; set; }
                 public DateTime Date_Column { get; set; }
-                public float Float_Column { get; set; }
+                public decimal Float_Column { get; set; }
                 public int Integer_Column { get; set; }
                 public int IntervalYearToMonth_Column { get; set; }
                 public TimeSpan InternalDayToSecond_Column { get; set; }
